requirements.txt
streamlit
openai
python-dotenv
PyPDF2
reportlab
fireworks-ai
pdfplumber
google-generativeai